+++ src/commands.js
@@ -5,7 +5,7 @@ module.exports = [
   { chained: true, name: 'validate', description: 'Checks your projects' },
   { chained: true, name: 'plan', description: 'Shows infra changes and saves in an executable plan' },
   { chained: true, name: 'backup', description: 'Backs up server state' },
-  { chained: true, name: 'launch', description: 'Launches virtual machines at a provider' },
+  { chained: true, name: 'infra', description: 'Launches virtual machines at a provider' },
   { chained: true, name: 'install', description: 'Runs Ansible to install software packages & configuration templates' },
   { chained: true, name: 'deploy', description: 'Upload your own application(s)' },
   { chained: true, name: 'restart', description: 'Restart your own application(s) and its dependencies' },